Refactor parser package
Move all JSON parser classes into the parser package
