+++ src/main/java/tecna/Storage.java
@@ -17,6 +17,10 @@ import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 
+import tecna.parser.DeadlineParser;
+import tecna.parser.EventParser;
+import tecna.parser.ToDoParser;
+
 /**
  * Saves and loads the data generated while the chatbot is running.
  * A <code>filePath</code> is the path to the file saving app data.
