+++ src/main/java/tecna/parser/ToDoParser.java
@@ -1,6 +1,10 @@
-package tecna;
+package tecna.parser;
 
 import org.json.simple.JSONObject;
+
+import tecna.JsonLoadingException;
+import tecna.JsonLoadingExceptionType;
+import tecna.ToDo;
 
 /**
  * Parses the task of ToDo type from a JSONObject.
